requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
astunparse==1.6.3
attrs==25.4.0
audioread==3.1.0
banal==1.0.6
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
dataset==1.6.2
datasets==4.8.3
decorator==4.4.2
defusedxml==0.7.1
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
docling==2.73.1
docling-core==2.70.1
docling-ibm-models==3.12.0
docling-parse==4.7.3
easyocr==1.7.2
et_xmlfile==2.0.0
facenet-pytorch==2.6.0
Faker==40.11.0
fastapi==0.135.1
fer==22.4.0
ffmpeg-python==0.2.0
ffmpy==1.0.0
filelock==3.25.2
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
future==1.0.0
gast==0.7.0
google-auth==2.48.0
google-genai==1.64.0
google-pasta==0.2.0
gradio==6.9.0
gradio_client==2.3.0
greenlet==3.3.2
groovy==0.1.2
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
joblib==1.5.3
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.13.2
kiwisolver==1.5.0
latex2mathml==3.79.0
lazy-loader==0.5
libclang==18.1.1
librosa==0.11.0
llama_cpp_python==0.3.16
llvmlite==0.46.0
lxml==6.0.2
lz4==4.4.5
Mako==1.3.10
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
more-itertools==10.8.0
moviepy==1.0.3
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.2
mtcnn==1.0.0
multidict==6.7.1
multiprocess==0.70.19
namex==0.1.0
networkx==3.6.1
ninja==1.13.0
numba==0.64.0
numpy==1.26.4
omegaconf==2.3.0
openai-whisper==20250625
opencv-contrib-python==4.11.0.86
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.19.0
orjson==3.11.7
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pillow==10.2.0
pip==26.0.1
platformdirs==4.9.4
pluggy==1.6.0
polyfactory==3.3.0
pooch==1.9.0
portalocker==3.2.0
proglog==0.1.12
propcache==0.4.1
protobuf==7.34.0
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.13.1
pydub==0.25.1
Pygments==2.19.2
pylatexenc==2.10
PyMuPDF==1.27.2.2
pyparsing==3.3.2
pypdfium2==5.6.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
python-pptx==1.0.2
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
qdrant-client==1.17.1
rapidocr==3.7.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
semchunk==3.2.5
sentence-transformers==5.3.0
setuptools==80.10.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
SQLAlchemy==1.4.54
starlette==0.52.1
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
tensorflow==2.21.0
termcolor==3.3.0
tf_keras==2.21.0
threadpoolctl==3.6.0
tifffile==2026.3.3
tiktoken==0.12.0
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.3
transformers==4.57.6
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.21.2
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
webcolors==25.10.0
websockets==15.0.1
wheel==0.46.3
wrapt==2.1.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
